feat: Add stop button to stop all currently playing sounds
The code changes in this commit add a stop button to the Main component. This button allows the user to stop all sounds that are currently playing. The functionality is implemented using the new `soundStore.stopAllSounds` function. This feature addresses the need for a convenient way to stop all sounds with a single click.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>