feat: Reset hotkey in HotkeyPicker.vue
This commit adds a new function `resetHotkey` to the `HotkeyPicker.vue` component. This function resets the selected hotkey to null and saves the null value to the `ptt_hotkey` setting in the `settingsStore`. This feature addresses the need to provide a way for the user to reset the hotkey to its default value.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>