feat: add PTT hotkey functionality
The code changes in this commit add the functionality to send the Push-to-Talk (PTT) hotkey. When the PTT hotkey is pressed down, the `_pttHotkeyPress` function is called, which sends the hotkey using the RobotJS API. This feature allows users to use a designated hotkey to activate their PTT when a sound is played.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>