feat: update HotkeyPicker to support multiple keys
The HotkeyPicker component has been updated to support multiple keys for the hotkey binding. Now, when a user presses a key, it will be added to the list of keys for the hotkey. The hotkey will be displayed as a combination of all the selected keys. If no keys are selected, the message "No Keybind Set" will be displayed.

This change improves the functionality of the HotkeyPicker component and allows for more flexibility in setting hotkeys.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>