feat: Fix bug preventing sound clip from playing when hotkey is pressed
The code changes in settings store addresses a bug where the sound clip would double-trigger when the hotkey was pressed. To fix this, the `disabled` state was added to the sound store to prevent playing sounds when the hotkey is pressed. This ensures that the sound clip is played only once when the hotkey is triggered.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>