feat: create InputTextNumber component
This commit adds a new component called InputTextNumber.vue. It provides an input field that allows users to enter numeric values. The component includes features such as incrementing and decrementing the value with arrow keys, as well as handling big steps and page up/down events. This component will be used in various parts of the application to improve user input for numeric values.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>