feat: add defaultVolume input to Settings.vue
This commit adds a new input field for the default volume in the Settings.vue component. Users can now enter a numeric value or use a range slider to select the default volume for the application. This enhancement improves the user experience by providing more control over the default volume setting.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>