fix: regedit not working in production
The code changes in package.json add an extraResources entry to copy the regedit/vbs folder from node_modules to the project directory. This change ensures that the required files for regedit are available for regedit.

In settings.js, the code changes update the vbCableIsInstalled function to use async/await syntax for improved readability. It also sets the external VBS location for regedit to "resources/regedit/vbs". This change ensures that the correct VBS location is used when using regedit to check if VBCable is installed.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>