style: update select-custom background colors for improved contrast
The code changes in SelectCustom.vue and main.css update the button and background colors to improve the contrast. The "--alt-bg-color" variable is now used for the background color, and the "--top-toolbar-color" variable is used for the button color. This change ensures better visibility and enhances the overall design consistency.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>