style: improve focus visibility for play button on SoundEditor component
This change improves the focus visibility for the play button in SoundEditor.vue. It fixes the issue where the button would falsely trigger focus events when the app sends a ptt_hotkey. By using the hard-coded focusVisible class, the focus outline is now consistent and improves the visibility and user experience when interacting with the SoundEditor.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>