feat: add tags to search functionality in Settings store
This commit enhances the SoundToolbar search functionality by allowing users to search for sounds based on tags, in addition to titles. The `soundsFiltered` getter in `settings.ts` has been modified to include the `tags` property in the search criteria. This change enables users to find sounds more accurately based on their tags, not just what is displayed on the button title.

Note: This commit builds upon previous commits that added tag-input functionality, created the TagInput component, and added the tags field to the SoundEditor component and Sound interface.
Signed-off-by: Jared Meadows <meadowsjared@gmail.com>