feat: update Soundboard component to use Transition for right sidebar
This refactor updates the Soundboard component to use the Vue Transition component for animating the right sidebar. The right sidebar now slides in and out smoothly when the currentEditingSound is not null. This change improves the user experience and adds a visual cue for the presence of the SoundEditor component.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>