style: update HotkeyPicker.vue and SoundToolbar to handle smaller sizes better
This commit updates the layout of the button-group in the HotkeyPicker.vue component. The flex-wrap property is added to allow the buttons to wrap to the next line when there is not enough horizontal space. This change improves the visual presentation of the buttons and ensures they are not cut off when the screen size is small.

In addition, the toolbar in SoundToolbar has gap and flex-wrap: wrap; added to assist in smaller layouts.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>