style: add minWidth and minHeight to window size in main.js
This commit adds the `minWidth` and `minHeight` properties to the window size in the `createWindow` function of the `main.js` file. By setting these properties, we ensure that the window cannot be resized below a certain width and height, improving the user experience and preventing content from being cut off.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>