refactor: improve search functionality in useSettingsStore
This refactor improves the search functionality in the useSettingsStore by enhancing the logic for filtering sounds based on the search text. The code now compares the words from the title and tags to the words from the search text, allowing for more accurate search results. Additionally, the code now handles uppercase tags correctly by converting them to lowercase before comparison. This change enhances the user experience and ensures that all relevant sounds are displayed in the search results.

Note: when it checks if searchWord contains the titleWord, we check if the titleWord.length > 1, because otherwise the search for "asdf" would match "lets a go".
It should be noted, it will still check if the title includes "asdf", therefore it should be good enough for now

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>