refactor: add function to add output device in Settings.vue
This refactor adds a new function, addOutputDevice, in the Settings.vue component. The function is called when an option is selected in the output device dropdown. It takes the selected device ID as a parameter and updates the outputDevices array accordingly. This change improves the functionality of the Settings page by allowing us to easily add a device elsewhere

Note: this is mainly because we want to be able to call this from downloadVBCable()
Signed-off-by: Jared Meadows <meadowsjared@gmail.com>