build: add @slosk/sudo-prompt dependency
This commit adds the "@slosk/sudo-prompt" npm dependency to the project. This package is required for performing privileged operations, such as running the installer as an administrator. By adding this dependency, we ensure that the necessary functionality is available for the application.  Previously the user needed to run "Pulse Panel" as an administrator in order to install VB-Cable.

Now that shouldn't be necessary anymore, as using this dependency, the app will attempt to run the installer as administrator

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>