chore: update downloadVBCable function to accept appName parameter
This commit updates the `downloadVBCable` function in the `electron-window.d.ts`, `main.js`, and `preload.js` files to accept an `appName` parameter. This change allows for more the Admin request dialog can show the appName, when in downloading VB-Cable. The function is modified in all three files to pass the `appName` parameter to the `downloadVBCable` and `runSetupAndCleanup` function in the `settings.js` file.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>