refactor: update VB-Cable installation process
This commit refactors the installation process for VB-Cable. It updates the `downloadVBCable` function in the `Settings.vue` component to accept an `appName` parameter. The function now checks if VB-Cable is already installed by verifying the existence of the `C:\\Program Files\\VB\\CABLE\\VBCABLE_ControlPanel.exe` file. If it is not installed, the function runs the installer as an administrator using the `@slosk/sudo-prompt` dependency. Additionally, the function updates the output devices after the installation to include the newly added device. These changes improve the user experience by simplifying the installation process and ensuring that VB-Cable is correctly installed.

refactor: the showVBCableMessage to use animation instead of setTimeout
Signed-off-by: Jared Meadows <meadowsjared@gmail.com>