build: update npm dependencies
Update the npm dependencies in the package.json file. This includes updating the versions of various packages such as vue, vue-inline-svg, @types/node, @typescript-eslint/parser, eslint, globals, vite, and pulse-panel. These updates ensure that the project is using the latest stable versions of the dependencies, which can improve performance, security, and compatibility.

Signed-off-by: Jared Meadows <meadowsjared@gmail.com>